feat(orcid_id): add class for handling orcid_id objects
